Remove unboxing

+++ doma-core/src/main/java/org/seasar/doma/wrapper/ByteWrapper.java
@@ -22,7 +22,7 @@ public class ByteWrapper extends AbstractWrapper<Byte> implements NumberWrapper<
   public void increment() {
     Byte value = doGet();
     if (value != null) {
-      doSet((byte) (value.byteValue() + 1));
+      doSet((byte) (value + 1));
     }
   }
 
@@ -30,7 +30,7 @@ public class ByteWrapper extends AbstractWrapper<Byte> implements NumberWrapper<
   public void decrement() {
     Byte value = doGet();
     if (value != null) {
-      doSet((byte) (value.byteValue() - 1));
+      doSet((byte) (value - 1));
     }
   }
